requirements_old_pande.txt
absl-py==2.4.0
accelerate==1.13.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bitsandbytes==0.49.2
brotli==1.2.0
ccimport==0.4.4
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
cumesh @ file:///tmp/extensions/CuMesh
cumm-cu118==0.7.11
cycler==0.12.1
deepspeed==0.18.8
easydict==1.13
einops==0.8.2
exceptiongroup==1.3.1
fastapi==0.133.1
ffmpy==1.0.0
filelock==3.24.3
fire==0.7.1
flash-attn==2.7.3
flatbuffers==25.12.19
flex_gemm @ file:///tmp/extensions/FlexGEMM
fonttools==4.62.1
fsspec==2026.2.0
glcontext==3.0.0
gradio==6.6.0
gradio_client==2.1.0
groovy==0.1.2
grpcio==1.78.1
h11==0.16.0
hf-xet==1.4.2
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kornia==0.8.2
kornia_rs==0.1.10
lark==1.3.1
lazy-loader==0.5
llvmlite==0.46.0
lpips==0.1.4
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
moderngl==5.12.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
ninja==1.13.0
numba==0.64.0
numpy==2.2.6
nvdiffrast @ file:///tmp/extensions/nvdiffrast
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.4.127
o_voxel @ file:///tmp/extensions/o-voxel
onnxruntime-gpu==1.23.2
opencv-python-headless==4.13.0.92
orjson==3.11.7
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pandas==2.3.3
pccm==0.4.16
pillow==10.2.0
Pillow-SIMD==9.5.0.post2
platformdirs==4.9.4
plyfile==1.1.3
pooch==1.9.0
portalocker==3.2.0
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pybind11==3.0.2
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyMatting==1.1.15
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
rembg==2.0.69
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
spconv-cu118==2.3.8
starlette==0.52.1
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
tifffile==2025.5.10
timm==1.0.25
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.3
transformers==5.3.0
trimesh==4.11.2
triton==3.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
utils3d @ git+https://github.com/EasternJournalist/utils3d.git@9a4eb15e4021b67b12c460c7057d642626897ec8
uvicorn==0.41.0
Werkzeug==3.1.6
xformers==0.0.29.post3
zstandard==0.25.0
